requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.1.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.4.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.25.1
shortuuid==1.0.1
sqlparse==0.4.1
urllib3==1.26.3
